data_analysis/requirements.txt
# Requirements for python code (outside of notebooks). Keep them sorted.
# Move requirements already present in pandas-base to installed_requirements.txt (unless you need to cap them).
# TODO(cyrille): Drop once this is the case in pandas-base.
airtable>=0.4.8
airtablemock==0.0.11
awscli==1.19.73
boto3==1.17.73
demjson==2.2.4
dnspython==2.1.0
elasticsearch==7.12.1
google-api-python-client==2.4.0
httplib2==0.19.1
js2py==0.71
json5==0.9.5
mypy==0.812
mypy-protobuf==2.4
nbformat==5.1.3
nose-terse-output==1.1.0
# For pdf2txt.py in Makefile
pdfminer==20191125
polib==1.1.1
protobuf==3.17.0
pycodestyle==2.7.0
pylint-doc-spacing==0.0.4
pylint-import-modules==0.2.0
pylint-quotes==0.2.1
pymongo-stubs==0.1.0
python-emploi-store==0.10.2
requests-aws4auth==1.0.1
requests_mock==1.9.2
runipy==0.1.5
sentry-sdk==1.1.0
termcolor==1.1.0
typeform==1.1.0
typing_extensions==3.10.0.0
unidecode==1.2.0
